requirements-p39.txt
absl-py==0.12.0
addict==2.4.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
autoscrape
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cityscapesScripts @ git+https://github.com/mcordts/cityscapesScripts.git@aeb7b82531f86185ce287705be28f452ba3ddbb8
cloudpickle==1.6.0
coloredlogs==15.0
colour==0.1.5
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
-e git+https://github.com/facebookresearch/detectron2.git@c1689c7d7b1c1af0157cd90b84f296b7b2ce2124#egg=detectron2
docopt==0.6.2
easydict==1.7
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
fvcore==0.1.5.post20210423
gast==0.4.0
gdown==3.12.2
glcontext==2.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
html5lib==1.1
humanfriendly==9.1
idna==2.10
imagecorruptions==1.1.2
imageio==2.9.0
iopath==0.1.8
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.5.0
librosa==0.8.0
llvmlite==0.36.0
lxml==4.6.3
magickwand==0.2
manimgl==1.0.0
ManimPango==0.2.5.post0
mapbox-earcut==0.12.10
Markdown==3.3.4
matplotlib==3.4.1
mmcv==1.3.1
moderngl==5.6.4
moderngl-window==2.3.0
mpmath==1.2.1
multipledispatch==0.6.0
networkx==2.5.1
numba==0.53.1
numpy
oauthlib==3.1.0
olefile
omegaconf==2.1.0.dev22
onnxruntime-gpu==1.7.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@118ffaba478b5d45925e88d504625b4123f3160a
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow
plyvel==1.3.0
pooch==1.3.0
portalocker==2.3.0
progressbar==2.5
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
pydub==0.25.1
pyglet==1.5.16
Pygments==2.8.1
PyOpenGL==3.1.5
pyparsing==2.4.7
pyquaternion==0.9.9
pyrr==0.10.3
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
screeninfo==0.6.7
selenium==3.141.0
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
sympy==1.8
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow==2.5.0rc1
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.9.1
tqdm==4.60.0
traitlets==5.0.5
typing==3.7.4.3
typing-extensions
urllib3==1.26.4
validators==0.18.2
Wand==0.6.6
warcio==1.7.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
yapf==0.31.0
